chg: move runPuppeteerCheck, rename puppeteer2/
