+++ src/lib/puppeteer/index.ts
@@ -3,3 +3,4 @@ export * from './page';
 export * from './types';
 export * from './metrics';
 export * from './pageStructure';
+export * from './runCheck';
